requirements.txt
-i https://pypi.org/simple
blinker==1.8.2; python_version >= '3.8'
click==8.1.7; python_version >= '3.7'
colorama==0.4.6; platform_system == 'Windows'
flask==3.0.3; python_version >= '3.8'
flask-mysql==1.5.2
flask-mysqldb==2.0.0; python_version >= '3.8'
flask-wtf==1.2.1; python_version >= '3.8'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
markupsafe==2.1.5; python_version >= '3.7'
mysqlclient==2.2.4; python_version >= '3.8'
pymysql==1.1.1; python_version >= '3.7'
python-dotenv==1.0.1; python_version >= '3.8'
werkzeug==3.0.4; python_version >= '3.8'
wtforms==3.1.2; python_version >= '3.8'
